Rename --init flag to --setup
Renames the configuration wizard flag from --init to --setup for better clarity. This includes renaming the init module to setup and updating all references throughout the codebase.
